fix for creating new Sortable with SortableJS 1.5.1
SortableJS 1.5.1 has a create() function for creating a new Sortable object. It was causing some TypeErrors as new Sortable() was not a constructor anymore.
This changes the implementation to create a new Sortable object using the create function instead.
Also, updates de README file for mapping correctly the minify version of sortablejs in SystemJS configuration.

Fixes #42